Refactor ampliconFilters to ampliconParams
Update API of sv_deletion_exp -- add 'aview' argument.  The aview
argument allows one to pass an AlignmentViews object. This is needed for
the LOOCV analysis since the file paths of the alignments are in a
different project directory than provided by the dirs argument.  (An
alternative that would not require changing the API would be to use
symbolic links in the new project directory).