getImproperAlignmentPairs does NOT save to disk
Prior to v0.0.4, getImproperAlignmentPairs saved the result to
improperPaths(object).  Inadvertently, the function
tag_sequenceExperiment in the svrearrange package overwrote these files
when the function tag_sequenceExperiment was called.  With the update,
tag_sequenceExperiment will only write to dp["3reads"] and the improper
alignments directory (dp["0improper"]) will not be touched.