Adjust blat score by length of tag sequence
When the tags are 125bp, a perfect match would have a score of 125 by
blat.  Further, only alignment sizes of 80 - 120 bps were passing QC.
With this update, the scores are adjusted by size and the acceptable
size of the alignment to the target genome is a window around the tag size.