check whether to compare fold-change to larger region
The germlineFilters function compares the segment mean to a larger neighborhood
of segments.  If the fold-change of an individual segment isn't substantially
reduced relative to the neighborhood, the segment is filtered. This commit
checks whether the neighborhood given the bins in the pviews object is
substantially larger than the total width of the cnv bins.  This is needed for
unit tests -- otherwise, the unit test ranges always get filtered.