Fix bug in initialize_graph
initialize_graph2 was modifying an element of the 'preprocess' list, but the list
was not returned to the environment of the calling function.  This commit adds a
new function amplified_segments that is called prior to initialize_graph2.  The
unit test for initalize_graph in test_amplicon_regressions was updated to
accomodate this change.